задание5
